Fix client SSL certificate and key support
When a client SSL certificate and key are defined in the Chef configuration, the contents of the files need to be read and passed to Spice's connection_options hash as OpenSSL objects (X509::Certificate and PKey).

Bumped the client gem version and cookbook version to 1.2.3 to reflect this change (since the client was still at 1.2.1 but the cookbook was 1.2.2) and updated some URLs and documentation.